Update HOW TO UPDATE.TXT
To make updates:  
Open Updates.xml & Edit.  
Copy:  
```
  <update>  
    <version>1.1000</version>  
    <file>1.1000.exe</file>  
  </update>  
```
To:  
```
  <update>  
    <version>1.1001</version>  
    <file>1.1001.exe</file>  
  </update> 
``` 
So now it should looks like this:  
```
  <update>
    <version>1.1000</version>
    <file>1.1000.exe</file>
  </update>
  <update>
    <version>1.1001</version>
    <file>1.1001.exe</file>
  </update>
  <update>
    <version>1.1002</version>
    <file>1.1002.exe</file>
  </update>
```
Assuming you want to update the list up to 1.1002.  
Make sure that the filename is correct or it will throw 404 message.  
So Application will download the next version 1.1001  
You can add as many updates as you want until 99999999.9999999 if you like.